+++ src/CMakeLists.txt
@@ -4,11 +4,16 @@ include_directories(
   ${DEPENDS_DIR}/libsnark/depends/libff
   ${DEPENDS_DIR}/libsnark/depends/libfqfft)
 
+add_library(ethsnarks_common STATIC export.cpp import.cpp stubs.cpp utils.cpp)
 
-add_library(miximus SHARED mod/miximus.cpp)
-target_link_libraries(miximus snark)
+add_library(miximus_objs OBJECT mod/miximus.cpp)
+
+add_library(miximus_static STATIC $<TARGET_OBJECTS:miximus_objs>)
+target_link_libraries(miximus_static snark)
+
+add_library(miximus SHARED $<TARGET_OBJECTS:miximus_objs>)
+target_link_libraries(miximus miximus_static)
 set_property(TARGET miximus PROPERTY POSITION_INDEPENDENT_CODE ON)
-
 
 add_library(hashpreimage SHARED mod/hashpreimage.cpp)
 target_link_libraries(hashpreimage snark)
@@ -16,11 +21,11 @@ set_property(TARGET hashpreimage PROPERTY POSITION_INDEPENDENT_CODE ON)
 
 
 add_executable(miximus_cli miximus_cli.cpp)
-target_link_libraries(miximus_cli snark)
+target_link_libraries(miximus_cli miximus_static)
 
 
 add_executable(hashpreimage_cli hashpreimage_cli.cpp)
-target_link_libraries(hashpreimage_cli snark)
+target_link_libraries(hashpreimage_cli snark ethsnarks_common)
 
 
 add_executable(verify verify.cpp)
@@ -40,7 +45,7 @@ target_link_libraries(test_shamir_poly snark)
 
 
 add_executable(test_sha256_full_gadget test/test_sha256_full_gadget.cpp)
-target_link_libraries(test_sha256_full_gadget snark)
+target_link_libraries(test_sha256_full_gadget snark ethsnarks_common)
 
 
 add_executable(test_proof_raw2json test/test_proof_raw2json.cpp)
@@ -48,11 +53,11 @@ target_link_libraries(test_proof_raw2json snark)
 
 
 add_executable(test_field_packing test/test_field_packing.cpp)
-target_link_libraries(test_field_packing snark)
+target_link_libraries(test_field_packing snark ethsnarks_common)
 
 
 add_executable(test_hashpreimage test/test_hashpreimage.cpp)
-target_link_libraries(test_hashpreimage snark)
+target_link_libraries(test_hashpreimage snark ethsnarks_common)
 
 
 add_executable(test_one_of_n test/test_one_of_n.cpp)
