+++ src/test/test_importexport.cpp
@@ -1,5 +1,5 @@
 #include "import.cpp"
-#include "export.cpp"
+#include "export.hpp"
 
 #include <sstream>
