+++ src/test/test_one_of_n.cpp
@@ -7,7 +7,7 @@
 
 #include "gadgets/one_of_n.cpp"
 
-#include "export.cpp"
+#include "export.hpp"
 #include "import.cpp"
