+++ src/test/test_proof_raw2json.cpp
@@ -1,6 +1,6 @@
-#include "export.cpp"
+#include "export.hpp"
 #include "import.cpp"
-#include "utils.cpp"
+#include "utils.hpp"
 
 
 using std::cout;
